Added Android lock/unlock feature
