+++ lib/devices/android/android-controller.js
@@ -339,7 +339,13 @@ androidController.postDismissAlert = function (cb) {
 };
 
 androidController.lock = function (secs, cb) {
-  cb(new NotYetImplementedError(), null);
+  this.adb.lock(function (err) {
+    if (err) return cb(err);
+    cb(null, {
+      status: status.codes.Success.code
+      , value: null
+    });
+  });
 };
 
 androidController.isLocked = function (cb) {
@@ -347,9 +353,69 @@ androidController.isLocked = function (cb) {
     if (err) return cb(err);
     cb(null, {
       status: status.codes.Success.code
-    , value: isLocked
+      , value: isLocked
     });
   });
+};
+
+androidController.pushUnlock = function (cb) {
+  logger.debug("Pushing unlock helper app to device...");
+  var unlockPath = path.resolve(__dirname, "..", "..", "..", "build",
+      "unlock_apk", "unlock_apk-debug.apk");
+  fs.stat(unlockPath, function (err) {
+    if (err) {
+      cb(new Error("Could not find unlock.apk; please run " +
+          "'reset.sh --android' to build it."));
+    } else {
+      this.adb.install(unlockPath, false, cb);
+    }
+  }.bind(this));
+};
+
+androidController.unlock = function (cb) {
+  this.adb.isScreenLocked(function (err, isLocked) {
+    if (err) return cb(err);
+    if (isLocked) {
+      logger.info("Unlocking screen");
+      var timeoutMs = 10000;
+      var start = Date.now();
+      var unlockAndCheck = function () {
+        logger.debug("Screen is locked, trying to unlock");
+        var onStart = function (err) {
+          if (err) return cb(err);
+          this.adb.isScreenLocked(function (err, isLocked) {
+            if (err) return cb(err);
+            if (!isLocked) {
+              logger.debug("Screen is unlocked, continuing");
+              return cb(null, {
+                status: status.codes.Success.code
+                , value: null
+              });
+            }
+            if ((Date.now() - timeoutMs) > start) {
+              return cb(new Error("Screen did not unlock"));
+            } else {
+              setTimeout(unlockAndCheck, 1000);
+            }
+          }.bind(this));
+        }.bind(this);
+        this.adb.startApp({
+          pkg: "io.appium.unlock",
+          activity: ".Unlock",
+          action: "android.intent.action.MAIN",
+          category: "android.intent.category.LAUNCHER",
+          flags: "0x10200000"
+        }, onStart);
+      }.bind(this);
+      unlockAndCheck();
+    } else {
+      logger.debug('Screen already unlocked, continuing.');
+      return cb(null, {
+        status: status.codes.Success.code
+        , value: null
+      });
+    }
+  }.bind(this));
 };
 
 androidController.equalsWebElement = function (cb) {
