+++ lib/devices/android/android.js
@@ -120,7 +120,7 @@ Android.prototype.start = function (cb, onDie) {
     this.pushUnlock.bind(this),
     this.uiautomator.start.bind(this.uiautomator),
     this.wakeUp.bind(this),
-    this.unlockScreen.bind(this),
+    this.unlock.bind(this),
     this.getDataDir.bind(this),
     this.setupCompressedLayoutHierarchy.bind(this),
     this.startAppUnderTest.bind(this),
