+++ lib/devices/android/chrome.js
@@ -113,7 +113,7 @@ ChromeAndroid.prototype.prepareChromedriver = function (cb) {
 ChromeAndroid.prototype.unlock = function (cb) {
   this.pushUnlock(function (err) {
     if (err) return cb(err);
-    this.unlockScreen(cb);
+    this.unlock(cb);
   }.bind(this));
 };
