+++ lib/devices/android/selendroid.js
@@ -77,6 +77,9 @@ Selendroid.prototype.init = function () {
 Selendroid.prototype.getSettings = deviceCommon.getSettings;
 Selendroid.prototype.updateSettings = deviceCommon.updateSettings;
 
+Selendroid.prototype.pushUnlock = androidController.pushUnlock;
+Selendroid.prototype.unlock = androidController.unlock;
+
 _.extend(Selendroid.prototype, androidCommon);
 Selendroid.prototype._deviceConfigure = Device.prototype.configure;
 Selendroid.prototype._setAndroidArgs = androidCommon.setAndroidArgs;
@@ -173,7 +176,7 @@ Selendroid.prototype.start = function (cb) {
     this.initUnicode.bind(this),
     this.pushSettingsApp.bind(this),
     this.pushUnlock.bind(this),
-    this.unlockScreen.bind(this),
+    this.unlock.bind(this),
     this.pushSelendroid.bind(this),
     this.waitForServer.bind(this)
   ], function (err) {
