+++ lib/server/controller.js
@@ -218,6 +218,14 @@ exports.lock = function (req, res) {
   }
 };
 
+exports.unlock = function (req, res) {
+  if (req.device.unlock) {
+    req.device.unlock(getResponseHandler(req, res));
+  } else {
+    notYetImplemented(req, res);
+  }
+};
+
 exports.isLocked = function (req, res) {
   req.device.isLocked(getResponseHandler(req, res));
 };
