+++ lib/server/routing.js
@@ -104,6 +104,7 @@ module.exports = function (appium) {
   // appium-specific extensions to JSONWP
   rest.post('/wd/hub/session/:sessionId?/appium/device/shake', controller.mobileShake);
   rest.post('/wd/hub/session/:sessionId?/appium/device/lock', controller.lock);
+  rest.post('/wd/hub/session/:sessionId?/appium/device/unlock', controller.unlock);
   rest.post('/wd/hub/session/:sessionId?/appium/device/is_locked', controller.isLocked);
   rest.post('/wd/hub/session/:sessionId?/appium/device/press_keycode', controller.pressKeyCode);
   rest.post('/wd/hub/session/:sessionId?/appium/device/long_press_keycode', controller.longPressKeyCode);
