+++ package.json
@@ -119,7 +119,7 @@
     "socks": "~0.0.1",
     "underscore-cli": "~0.2.17",
     "unorm": "~1.3.3",
-    "wd": "~0.3.7",
+    "wd": "~0.3.8",
     "wd-bridge": "~0.0.1",
     "yiewd": "~0.5.0"
   }
